Change seasons vs season references to be uniform
