+++ lib/seasons_collection.rb
